"changed all index ints to unsigned ints and trash delete"

+++ UDBlockCommentAutomaton.cpp
@@ -22,7 +22,7 @@ void UDBlockCommentAutomaton::S1(const std::string& input) {
 }
 
 void UDBlockCommentAutomaton::S2(const std::string& input) {
-    if(input[index] != '|' && index != input.size()-1 )
+    if(input[index] != '|' && (unsigned int)index != input.size()-1 )
     {
         inputRead++;
         index++;
